backend/requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
numpy==1.22.3
pandas==1.4.2
scikit-learn==1.3.0
opensearch-py
boto3
mysql-connector-python
flask-jwt-extended==4.4.4
PyJWT==2.6.0
bcrypt==4.0.1
flask_mail==0.9.1
stripe==5.5.0
google-auth==2.8.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
python-dotenv==1.0.0
ray==2.5.1
openai==1.6.1
google-api-python-client==2.95.0
flask_socketio
Werkzeug==2.2.2
twilio
cryptography
fpdf
anthropic
PyPDF2
ratelimit
langchain
sec_api
pycryptodome==3.20.0
pydantic==1.10.12